fix(webapp): remove cdp/ -> scoops/ back-edges
Move the CDP response chunking helpers into @slicc/shared-ts next to the
wire union they operate on, and import TrayTargetEntry from shared-ts
directly instead of via the scoops/ re-export. cdp/ now has zero imports
from scoops/.

Closes #1950
