+++ packages/shared-ts/CLAUDE.md
@@ -12,6 +12,7 @@ Platform-agnostic primitives shared across `@slicc/webapp`, `@slicc/node-server`
 - `tray-signaling.ts` — tray signaling wire contract (leader↔worker control messages, follower HTTP bootstrap request/response shapes, bootstrap retry constants). Single source of truth for `@slicc/webapp` and `@slicc/cloudflare-worker`; the iOS follower mirrors a subset in `packages/ios-app/SliccTrayKit/Models/TrayTypes.swift`. Worker-internal persisted state (`TrayBootstrapRecord`, `TrayRecord`) stays in `packages/cloudflare-worker/src/shared.ts`.
 - `agent-wire-types.ts` — agent/chat wire payload types (`AgentEvent`, `ChatMessage`, `ToolCall`, `MessageAttachment`, `LickEvent`) embedded in the tray sync protocol. Types only; the producing/consuming behavior stays in `@slicc/webapp`, which re-exports them from their original layer-local modules.
 - `tray-sync-protocol.ts` — tray sync data-channel wire contract (`LeaderToFollowerMessage` / `FollowerToLeaderMessage` unions, summary/target/FS types, `TRAY_SYNC_PROTOCOL_VERSION`, `unhandledProtocolMessage`). Partially mirrored by `packages/ios-app/SliccTrayKit/Models/SyncProtocol.swift` and enforced by the golden-fixture corpus (`packages/webapp/src/scoops/tray-sync-protocol-corpus.ts`). The `TraySyncChannel` runtime stays in `@slicc/webapp` `scoops/tray-sync-protocol.ts`.
+  - Also holds the CDP response chunking helpers (`sendCDPResponse`, `reassembleCDPResponse`, `CDP_CHUNK_THRESHOLD`). They are pure over the `cdp.response` message variant, so they live next to the wire union rather than in the webapp `scoops/` layer, where `cdp/`-layer callers could only reach them through a back-edge.
   - Also holds the transport chunk frame (`TrayChunkFrame`, `TRAY_CHUNK_FRAME_TYPE`, `isTrayChunkFrame`) and its bounds (`TRAY_DEFAULT_MAX_MESSAGE_BYTES`, `TRAY_MAX_MESSAGE_BYTES`, `TRAY_SEND_HIGH_WATER_BYTES`). A frame sits **below** the message unions — senders split an oversize message into frames and receivers reassemble before decoding the union — so it has no corpus fixture and no per-variant iOS expectation. Mirrored by `packages/slicc-cli/internal/protocol` (`ChunkFrame`) and `packages/ios-app/SliccTrayKit/Models/TrayChunkFraming.swift`.
 - `bridge-protocol.ts` — standalone/Electron thin-bridge launch contract: `slicc.bridge.v1.` subprotocol prefix, `bridge`/`bridgeToken` query params, `X-Bridge-Token` header, and the **single source of truth** for the hosted origins (`SLICC_HOSTED_ORIGIN`, `SLICC_STAGING_HUB_ORIGIN`) that every TS package imports instead of hardcoding. The `lint:hosted-origin` gate (`packages/dev-tools/tools/check-hosted-origin-literal.mjs`) prevents raw literals from re-appearing. Swift mirror: `packages/swift-server/Sources/Server/BridgeSecurity.swift`.
 - `sign-and-forward.ts` — S3 / Adobe da.live sign-and-forward orchestration (`executeS3SignAndForward` / `executeDaSignAndForward`). Validates an envelope, resolves credentials via an async `SecretGetter`, signs (S3) or attaches a Bearer token (DA), forwards, and returns a JSON-cloneable reply. Consumed by the webapp mount barrel, the node-server Express handlers (via a `SecretStore` adapter), and the extension service worker.
