+++ packages/webapp/src/scoops/tray-sync-protocol.ts
@@ -20,6 +20,7 @@ import type {
   TraySyncMessage,
 } from '@slicc/shared-ts';
 import {
+  CDP_CHUNK_THRESHOLD,
   isTrayChunkFrame,
   TRAY_CHUNK_FRAME_TYPE,
   TRAY_DEFAULT_MAX_MESSAGE_BYTES,
@@ -71,122 +72,6 @@ export {
 } from '@slicc/shared-ts';
 
 const log = createLogger('tray-sync');
-
-// ---------------------------------------------------------------------------
-// CDP response chunking helpers
-// ---------------------------------------------------------------------------
-
-/** Chunk size threshold in bytes — CDP responses larger than this are chunked. */
-export const CDP_CHUNK_THRESHOLD = 64 * 1024; // 64 KB
-
-/** Individual chunk size — smaller than threshold for safety margin. */
-const CDP_CHUNK_SIZE = 32 * 1024; // 32 KB
-
-/** Extract the CDP response message type from a union. */
-type CDPResponseMessage = Extract<TraySyncMessage, { type: 'cdp.response' }>;
-
-/**
- * Send a CDP response, automatically chunking if the serialized result exceeds CDP_CHUNK_THRESHOLD.
- * Returns true if all chunks were sent successfully, false if any send failed.
- */
-export function sendCDPResponse(
-  channel: { send(message: TraySyncMessage): boolean },
-  requestId: string,
-  result?: Record<string, unknown>,
-  error?: string
-): boolean {
-  // Error responses are always small — send directly
-  if (error || !result) {
-    return channel.send({ type: 'cdp.response', requestId, result, error } as CDPResponseMessage);
-  }
-
-  const serialized = JSON.stringify(result);
-  if (serialized.length <= CDP_CHUNK_THRESHOLD) {
-    // Small enough — send as a single message
-    return channel.send({ type: 'cdp.response', requestId, result } as CDPResponseMessage);
-  }
-
-  // Split the serialized result into chunks
-  const totalChunks = Math.ceil(serialized.length / CDP_CHUNK_SIZE);
-  let allSent = true;
-  for (let i = 0; i < totalChunks; i++) {
-    const chunkData = serialized.slice(i * CDP_CHUNK_SIZE, (i + 1) * CDP_CHUNK_SIZE);
-    const ok = channel.send({
-      type: 'cdp.response',
-      requestId,
-      chunkData,
-      chunkIndex: i,
-      totalChunks,
-    } as CDPResponseMessage);
-    if (!ok) {
-      allSent = false;
-      // Send an error response to unblock the requester (error messages are small, will fit)
-      channel.send({
-        type: 'cdp.response',
-        requestId,
-        error: `Failed to send CDP response chunk ${i}/${totalChunks} (response was ${serialized.length} bytes)`,
-      } as CDPResponseMessage);
-      break;
-    }
-  }
-  return allSent;
-}
-
-/**
- * Reassemble chunked CDP responses. Returns the parsed result when all chunks
- * have arrived, or null if still waiting for more chunks.
- *
- * @param buffers - shared buffer map, keyed by requestId
- * @param requestId - the request ID
- * @param message - the incoming cdp.response message
- * @returns { result, error } when complete, null when still accumulating
- */
-export function reassembleCDPResponse(
-  buffers: Map<string, { chunks: string[]; received: number; totalChunks: number }>,
-  message: CDPResponseMessage
-): { result?: Record<string, unknown>; error?: string } | null {
-  // Non-chunked response — return directly
-  if (message.chunkIndex === undefined || message.totalChunks === undefined) {
-    return { result: message.result, error: message.error };
-  }
-
-  // If this is an error during chunked transfer, abort and return error
-  if (message.error) {
-    buffers.delete(message.requestId);
-    return { error: message.error };
-  }
-
-  const requestId = message.requestId;
-  let buffer = buffers.get(requestId);
-  if (!buffer) {
-    buffer = {
-      chunks: new Array(message.totalChunks),
-      received: 0,
-      totalChunks: message.totalChunks,
-    };
-    buffers.set(requestId, buffer);
-  }
-
-  // Store the chunk (supports out-of-order delivery)
-  if (!buffer.chunks[message.chunkIndex]) {
-    buffer.chunks[message.chunkIndex] = message.chunkData!;
-    buffer.received++;
-  }
-
-  if (buffer.received >= buffer.totalChunks) {
-    buffers.delete(requestId);
-    try {
-      const result = JSON.parse(buffer.chunks.join('')) as Record<string, unknown>;
-      return { result };
-    } catch (err) {
-      return {
-        error: `Failed to reassemble CDP response: ${err instanceof Error ? err.message : String(err)}`,
-      };
-    }
-  }
-
-  return null; // Still waiting for more chunks
-}
 
 // ---------------------------------------------------------------------------
 // Snapshot chunking helpers
